fix(anvil): apply load state in memdb
